FOGL-6036: implementation of time difference
FOGL-6036: implementation of time difference

WatchDog routines moved to separate file
